lock diagram, set ports visibility:hidden and move custom node files to folder
