--- love/src/components/GeneralPurpose/Graph/DiamondNodeWidget.module.css
@@ -1,12 +1,0 @@
-.port {
-  width: 16px;
-  height: 16px;
-  z-index: 10;
-  background: rgba(0, 0, 0, 0.5);
-  border-radius: 8px;
-  cursor: pointer;
-}
-
-.port:hover {
-  background: rgba(0, 0, 0, 1);
-}
