+++ love/src/components/GeneralPurpose/Graph/Node/DiamondNodeFactory.jsx
