+++ love/src/components/GeneralPurpose/Graph/Node/DiamondNodeModel.jsx
