+++ love/src/components/GeneralPurpose/Graph/Node/DiamondPortModel.jsx
